Some unit tests added.

+++ app/src/main/java/com/tinmegali/springrestoauthandroidclient/models/ErrorHttp.java
@@ -10,7 +10,7 @@ public class ErrorHttp {
 
     @SerializedName("timestamp")
     @Expose
-    private Integer timestamp;
+    private Long timestamp;
     @SerializedName("status")
     @Expose
     private Integer status;
@@ -29,7 +29,7 @@ public class ErrorHttp {
      * @return
      * The timestamp
      */
-    public Integer getTimestamp() {
+    public Long getTimestamp() {
         return timestamp;
     }
 
@@ -38,7 +38,7 @@ public class ErrorHttp {
      * @param timestamp
      * The timestamp
      */
-    public void setTimestamp(Integer timestamp) {
+    public void setTimestamp(Long timestamp) {
         this.timestamp = timestamp;
     }
